Qx12xr

+++ FBOLinx.Web/FBOLinx.Web/ClientApp/src/app/pages/pricing-templates/pricing-templates-edit/pricing-templates-edit.component.ts
@@ -1,27 +1,29 @@
-import { Component, EventEmitter, Input, Output, OnInit, ViewChild } from '@angular/core';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
+import { Component, EventEmitter, Input, Output, ViewChild } from '@angular/core';
+import { MatDialog,  MAT_DIALOG_DATA } from '@angular/material';
 import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 
 //Services
 import { CustomermarginsService } from '../../../services/customermargins.service';
+import { FbopricesService } from '../../../services/fboprices.service';
 import { PricetiersService } from '../../../services/pricetiers.service';
 import { PricingtemplatesService } from '../../../services/pricingtemplates.service';
 import { SharedService } from '../../../layouts/shared-service';
 
 //Components
 import { DistributionWizardMainComponent } from '../../../shared/components/distribution-wizard/distribution-wizard-main/distribution-wizard-main.component';
+import { RichTextEditorComponent } from '@syncfusion/ej2-angular-richtexteditor';
 
 const BREADCRUMBS: any[] = [
     {
         title: 'Main',
-        link: '#/'
+        link: '#/default-layout'
     },
     {
-        title: 'Pricing Templates',
+        title: 'Margin Templates',
         link: '#/default-layout/pricing-templates'
     },
     {
-        title: 'Edit Pricing Template',
+        title: 'Edit Margin Template',
         link: ''
     }
 ];
@@ -41,17 +43,22 @@ export class PricingTemplatesEditComponent {
     @Output() savePricingTemplateClicked = new EventEmitter<any>();
     @Output() cancelPricingTemplateEditclicked = new EventEmitter<any>();
     @Input() pricingTemplate: any;
+    @ViewChild('typeRTE') rteObj: RichTextEditorComponent;
+    @ViewChild('typeEmail') rteEmail: RichTextEditorComponent;
 
     //Public Members
-    public pageTitle: string = 'Edit Pricing Template';
+    public focus: any = false;
+    public count: number = 0;
+    public pageTitle: string = 'Edit Margin Template';
     public breadcrumb: any[] = BREADCRUMBS;
     public marginTypeDataSource: Array<any> = [
         { text: 'Cost +', value: 0 },
-        { text: 'Retail -', value: 1 },
-        { text: 'Flat Fee', value: 2 },
-        { text: 'Inactive', value: 3 }
+        { text: 'Retail -', value: 1 }
+        //{ text: 'Flat Fee', value: 2 }
     ];
     public isSaving: boolean = false;
+    public currentPrice: any;
+    public product: any;
 
     /** pricing-templates-edit ctor */
     constructor(private route: ActivatedRoute,
@@ -59,29 +66,75 @@ export class PricingTemplatesEditComponent {
         private customerMarginsService: CustomermarginsService,
         private priceTiersService: PricetiersService,
         private pricingTemplatesService: PricingtemplatesService,
+        private fbopricesService: FbopricesService,
         private sharedService: SharedService,
         public distributionDialog: MatDialog    ) {
 
         this.sharedService.emitChange(this.pageTitle);
+        
+        
+    }
 
+    ngOnInit(): void {
         //Check for passed in id
         let id = this.route.snapshot.paramMap.get('id');
-        if (!id)
+        if (!id) {
+            this.loadCurrentPrice();
             return;
-        else {
+        } else {
             this._RequiresRouting = true;
-            this.pricingTemplatesService.get({oid: id}).subscribe((data: any) => {
+            this.pricingTemplatesService.get({ oid: id }).subscribe((data: any) => {
                 this.pricingTemplate = data;
-                this.customerMarginsService.getCustomerMarginsByPricingTemplateId(this.pricingTemplate.oid).subscribe((data: any) => {
-                    this.pricingTemplate.customerMargins = data;
-                });
+                this.customerMarginsService.getCustomerMarginsByPricingTemplateId(this.pricingTemplate.oid).subscribe(
+                    (data: any) => {
+                        this.pricingTemplate.customerMargins = data;
+
+                        this.loadCurrentPrice();
+                    });
             });
+        }
+    }
+
+    public disableToolbarNote() {
+        //this.rteObj.toolbarSettings.enable = false;
+
+    }
+
+    public enableToolbarNote() {
+        this.pricingTemplate.notes = this.rteObj.getText();
+        this.rteObj.toolbarSettings.enable = true;
+    }
+
+    public disableToolbarEmail() {
+        //this.rteEmail.toolbarSettings.enable = false;
+    }
+
+    public enableToolbarEmail() {
+        this.pricingTemplate.email = this.rteEmail.getText();
+        this.rteEmail.toolbarSettings.enable = true;
+    }
+
+    public onFocus() {
+        this.focus = true;
+    }
+
+    public outFocus() {
+        this.focus = false;
+        this.count = 0;
+    }
+
+
+    public onType(customerMargin, event) {
+        this.count = this.count + 1;
+        if (this.focus && this.count == 1) {
+            customerMargin.amount = event.key / 100;
         }
     }
 
     public savePricingTemplate() {
         this.isSaving = true;
-        //Update the pricing template first
+        
+        //Update the margin template first
         this.pricingTemplatesService.update(this.pricingTemplate).subscribe((data: any) => {
             this.saveCustomerMargins();
         });
@@ -96,13 +149,19 @@ export class PricingTemplatesEditComponent {
     }
 
     public deleteCustomerMargin(customerMargin) {
-        this.customerMarginsService.remove(customerMargin).subscribe((data: any) => {
-            this.priceTiersService.remove({ oid: customerMargin.priceTierId }).subscribe((data: any) => {
-                this.pricingTemplate.customerMargins.splice(
-                    this.pricingTemplate.customerMargins.indexOf(customerMargin),
-                    1);
+        if (customerMargin.oid == 0) {
+            this.pricingTemplate.customerMargins.splice(
+                this.pricingTemplate.customerMargins.indexOf(customerMargin),
+                1);
+        } else {
+            this.customerMarginsService.remove(customerMargin).subscribe((data: any) => {
+                this.priceTiersService.remove({ oid: customerMargin.priceTierId }).subscribe((data: any) => {
+                    this.pricingTemplate.customerMargins.splice(
+                        this.pricingTemplate.customerMargins.indexOf(customerMargin),
+                        1);
+                });
             });
-        });
+        }
     }
 
     public addCustomerMargin() {
@@ -112,40 +171,97 @@ export class PricingTemplatesEditComponent {
             priceTierId: 0,
             min: 1,
             max: 99999,
-            amount: 0
+            amount: 0,
+            itp: 0,
+            allin: 0
         };
         if (this.pricingTemplate.customerMargins.length > 0) {
             customerMargin.min =
                 this.pricingTemplate.customerMargins[this.pricingTemplate.customerMargins.length - 1].min + 250;
         }
         this.pricingTemplate.customerMargins.push(customerMargin);
-        this.fixCustomerMargins();
+       
+       // this.fixCustomerMargins();
     }
 
     public fixCustomerMargins() {
         for (let i in this.pricingTemplate.customerMargins) {
             let indexNumber = Number(i);
+            if (!this.pricingTemplate.customerMargins[indexNumber].min || this.pricingTemplate.customerMargins[indexNumber].min == '') {
+                if (indexNumber == 0)
+                    this.pricingTemplate.customerMargins[indexNumber].min = 1;
+                else
+                    this.pricingTemplate.customerMargins[indexNumber].min =
+                        (this.pricingTemplate.customerMargins[indexNumber - 1].min + 1);
+            }
+
+            if (indexNumber > 0 && this.pricingTemplate.customerMargins[indexNumber].min == this.pricingTemplate.customerMargins[indexNumber - 1].min) {
+                this.pricingTemplate.customerMargins[indexNumber].min =
+                    (this.pricingTemplate.customerMargins[indexNumber - 1].min + 1);
+            }
+                
             if (this.pricingTemplate.customerMargins.length > (indexNumber + 1) &&
                 this.pricingTemplate.customerMargins[indexNumber + 1].min > 0)
                 this.pricingTemplate.customerMargins[i].max = this.pricingTemplate.customerMargins[indexNumber + 1].min - 1;
             else
                 this.pricingTemplate.customerMargins[i].max = 99999;
+
+            this.calculateItpForMargin(this.pricingTemplate.customerMargins[indexNumber]);
+        }
+    }
+
+    public updateCustomerMargin(margin) {
+        
+        var jetACost = this.currentPrice.filter(item => item.product == 'JetA Cost')[0].price;
+        var jetARetail = this.currentPrice.filter(item => item.product == 'JetA Retail')[0].price;
+
+        if (this.pricingTemplate.marginType == 0) {
+            if (margin.min !=null && margin.amount !=null) {
+                margin.allin = jetACost + margin.amount;
+            }
+
+        }
+        else if (this.pricingTemplate.marginType == 1) {
+            if (margin.amount !=null && margin.min !=null) {
+                margin.allin = jetARetail - margin.amount;
+                if (margin.allin) {
+                    margin.itp = margin.allin - jetACost;
+                }
+            }
         }
     }
 
     public distributePricingTemplate() {
-        var pricingTemplates = [this.pricingTemplate];
+        var pricingTemplate = this.pricingTemplate;
         const dialogRef = this.distributionDialog.open(DistributionWizardMainComponent, {
             data: {
-                pricingTemplates: pricingTemplates,
+                pricingTemplate: pricingTemplate,
                 fboId: this.sharedService.currentUser.fboId,
                 groupId: this.sharedService.currentUser.groupId
-            }
+            },
+            disableClose: true
         });
 
         dialogRef.afterClosed().subscribe(result => {
 
         });
+    }
+
+    public customerMarginAmountChanged(customerMargin) {
+        var indexOfMargin = this.pricingTemplate.customerMargins.indexOf(customerMargin);
+        if (indexOfMargin > 0) {
+            var previousTier = this.pricingTemplate.customerMargins[indexOfMargin - 1];
+            if ((this.pricingTemplate.marginType == 0 || this.pricingTemplate.marginType == 2) && Math.abs(previousTier.amount) < Math.abs(customerMargin.amount))
+                customerMargin.amount = previousTier.amount + .01;
+            else if (this.pricingTemplate.marginType == 1 && Math.abs(previousTier.amount) > Math.abs(customerMargin.amount))
+                customerMargin.amount = previousTier.amount - .01;
+        }
+        this.calculateItpForMargin(customerMargin);
+        this.calculateAllInForMargin(customerMargin);
+    }
+
+    public marginTypeChange() {
+        this.loadCurrentPrice();
     }
 
     //Private Methods
@@ -162,5 +278,65 @@ export class PricingTemplatesEditComponent {
                     this.savePricingTemplateClicked.emit();
                 this.isSaving = false;
             });
+    }
+
+    private loadCurrentPrice() {
+        //this.fbopricesService.getFbopricesByFboIdAndProductCurrent(this.sharedService.currentUser.fboId, this.getCurrentProductForMarginType()).subscribe((data: any) => {
+        //    this.currentPrice = data;
+        //    for (let margin of this.pricingTemplate.customerMargins) {
+        //        this.calculateItpForMargin(margin);
+        //    }
+        //});
+
+        this.fbopricesService.getFbopricesByFboIdCurrent(this.sharedService.currentUser.fboId).subscribe((data: any) => {
+            this.currentPrice = data;
+            for (let margin of this.pricingTemplate.customerMargins) {
+                //this.calculateItpForMargin(margin);
+                this.updateCustomerMargin(margin);
+            }
+        });
+    }
+
+    private getCurrentProductForMarginType() {
+        if (this.pricingTemplate.marginType == 0)
+            return 'JetA Cost';
+        return 'JetA Retail';
+    }
+
+    private calculateItpForMargin(customerMargin) {
+        for (let m of this.currentPrice) {
+            if (this.pricingTemplate.marginType == 0 && this.getCurrentProductForMarginType() == m.product) {
+                customerMargin.itp = m ? m.price : 0 + Math.abs(customerMargin.amount);
+                return;
+            }
+            else if (this.pricingTemplate.marginType == 1 && this.getCurrentProductForMarginType() == m.product) {
+                customerMargin.itp = m ? m.price : 0 - Math.abs(customerMargin.amount);
+                return;
+            }
+            else if (this.pricingTemplate.marginType > 1) {
+                customerMargin.itp = Math.abs(customerMargin.amount);
+                return;
+            }
+        }
+    }
+
+    private calculateAllInForMargin(customerMargin) {
+        var jetACost = this.currentPrice.filter(item => item.product == 'JetA Cost')[0].price;
+        var jetARetail = this.currentPrice.filter(item => item.product == 'JetA Retail')[0].price;
+
+        if (this.pricingTemplate.marginType == 0) {
+            if (customerMargin.min != null && customerMargin.amount != null) {
+                customerMargin.allin = jetACost + customerMargin.amount;
+            }
+
+        }
+        else if (this.pricingTemplate.marginType == 1) {
+            if (customerMargin.amount != null && customerMargin.min != null) {
+                customerMargin.allin = jetARetail - customerMargin.amount;
+                if (customerMargin.allin) {
+                    customerMargin.itp = customerMargin.allin - jetACost;
+                }
+            }
+        }
     }
 }
